fixes #325
